backend/pyproject.toml
[tool.poetry]
name = "madgab-online"
version = "0.1.0"
description = ""
authors = ["Benjamin Greenawald <bgreenawald@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
simplejson = "^3.17.0"
apscheduler = "^3.6.3"
flask = "^1.1.1"
flask_cors = "^3.0.8"
flask_socketio = "^4.2.1"
numpy = "^1.18.1"
scipy = "^1.4.1"
pandas = "^0.25.3"
jsonschema = "^3.2.0"
gunicorn = {extras = ["eventlet"], version = "^20.0.4"}


[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.761"
flake8 = "^3.7.9"
flake8-black = "^0.1.1"
flake8-import-order = "^0.18.1"
flake8-bugbear = "^20.1.2"
flake8-bandit = "^2.1.2"
coverage = "^5.0.4"
pytest = "^5.4.1"
pytest-xdist = "^1.32.0"
python-dotenv = "^0.15.0"
requests = "^2.25.1"
tqdm = "^4.56.0"

[tool.black]
max-line-length=100

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
